+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 0.73.4 (2022-05-27)
+## [Unreleased]
+
+### Bug Fixes
+
+- **#1466** In LaTeX mode, doing a Select All (cmd+A), then delete would put the
+  mathfield in an inconsistent state
+- While in LaTeX mode, doing a Select All (cmd+A) with a partial command
+  followed by an auto-complete suggestion would render the mathfield
+  unresponsive
 
 ## 0.73.4 (2022-05-27)